refactor(datalab): ♻️ Move issue managers to separate modules in issue_manager subpackage
